+++ src/sender/buffer.rs
@@ -1,7 +1,7 @@
 use crate::sender::ProtocolEncoder;
 
 pub(crate) struct PulsedataBuffer<const S: usize> {
-    pub buf: [usize; S],
+    pub buf: [u32; S],
     pub offset: usize,
 }
 
@@ -17,22 +17,22 @@ impl<const S: usize> PulsedataBuffer<S> {
         self.offset = 0;
     }
 
-    pub fn load<SendProto: ProtocolEncoder<F>, const F: usize>(&mut self, c: &SendProto::Cmd) {
+    pub fn load<SendProto: ProtocolEncoder<F>, const F: u32>(&mut self, c: &SendProto::Cmd) {
         let len = SendProto::encode(c, &mut self.buf[self.offset..]);
         self.offset += len;
     }
 
-    pub fn get(&self, index: usize) -> Option<usize> {
+    pub fn get(&self, index: usize) -> Option<u32> {
         self.buf.get(index).cloned()
     }
 
-    pub fn buffer(&self) -> &[usize] {
+    pub fn buffer(&self) -> &[u32] {
         &self.buf[..self.offset]
     }
 }
 
 impl<'a, const S: usize> IntoIterator for &'a PulsedataBuffer<S> {
-    type Item = usize;
+    type Item = u32;
     type IntoIter = PulseIterator<'a>;
 
     fn into_iter(self) -> Self::IntoIter {
@@ -44,12 +44,12 @@ impl<'a, const S: usize> IntoIterator for &'a PulsedataBuffer<S> {
 }
 
 pub struct PulseIterator<'a> {
-    pulses: &'a [usize],
+    pulses: &'a [u32],
     pos: usize,
 }
 
 impl<'a> Iterator for PulseIterator<'a> {
-    type Item = usize;
+    type Item = u32;
 
     fn next(&mut self) -> Option<Self::Item> {
         if self.pos == self.pulses.len() {
