+++ src/sender/mod.rs
@@ -12,9 +12,9 @@ pub use buffer::*;
 pub use hal::*;
 pub use senders::*;
 
-pub trait ProtocolEncoder<const FREQ: usize>: Protocol {
+pub trait ProtocolEncoder<const FREQ: u32>: Protocol {
     type EncoderData;
     const DATA: Self::EncoderData;
 
-    fn encode(cmd: &Self::Cmd, buf: &mut [usize]) -> usize;
+    fn encode(cmd: &Self::Cmd, buf: &mut [u32]) -> usize;
 }
